+++ nodes/string/string.trim/index.js
@@ -6,14 +6,6 @@ module.exports = (NODE) => {
   const resultOut = NODE.getOutputByName('result');
   resultOut.on('trigger', async (conn, state) => {
     const strs = await stringsIn.getValues(state);
-    let result = '';
-    for (let i = 0; i < strs.length; i += 1) {
-      const str = strs[i];
-      if (typeof str === 'string') {
-        result += str.trim();
-      }
-    }
-
-    return result;
+    return strs.map(str => str.trim())
   });
 };